requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.3.0
stripe==12.3.0
channels==4.2.2
asgiref==3.9.0
django-summernote==0.8.20.0
WeasyPrint==61.2 
channels-redis>=4.1.0 
daphne>=4.0.0 
drf-yasg>=1.21.7 